Release of version 0.3 (#76)
The jump from version 0.2 to 0.3 is a huge one. A lot happened in the background. The two most significant changes are the test pipeline and also the removal of the underlying API and the use of the official lotrointerface xml API.

Otherwise the release consists several different bugfixes and a convenient backup functionality